Inclusão de Regras de Negócio e Tratamento de Exceções

+++ src/main/java/com/epcafes/service/CustoFixoService.java
@@ -7,7 +7,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.epcafes.exception.BusinessException;
 import com.epcafes.model.CustoFixo;
+import com.epcafes.model.DespesaCustoFixo;
 import com.epcafes.repository.CustoFixoRepository;
 
 @Service
@@ -15,6 +17,9 @@ public class CustoFixoService {
 	
 	@Autowired
 	private CustoFixoRepository custoFixoRepository;
+	
+	@Autowired
+	private DespesaCustoFixoService despesaCustoFixoService;
 	
 	@Transactional
 	public CustoFixo salvar(CustoFixo custoFixo) {
@@ -33,7 +38,14 @@ public class CustoFixoService {
 	}
 	
 	@Transactional
-	public void excluir(Long id) {
+	public void excluir(Long id) throws BusinessException {
+		
+		Optional<CustoFixo> custoFixo = custoFixoRepository.findById(id);
+		
+		List<DespesaCustoFixo> despesas = despesaCustoFixoService.listarDespesasDeUmCustoFixo(custoFixo.get());
+		
+		if(!despesas.isEmpty())
+			throw new BusinessException("", "Não é possivel excluir o Custo Fixo " + custoFixo.get().getNome() + " pois possui despesas cadastradas!");
 		
 		custoFixoRepository.deleteById(id);
 	}
